refactor: integration과 unit으로 테스트 분리
